prd/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
streamlit>=1.33.0
fastmcp>=0.1.0
chromadb>=0.4.24
anthropic>=0.25.0
google-genai>=1.0.0
pymupdf>=1.24.0
python-docx>=1.1.0
pytesseract>=0.3.10
Pillow>=10.3.0
presidio-analyzer>=2.2.354
presidio-anonymizer>=2.2.354
pyyaml>=6.0.1
python-dotenv>=1.0.1
python-multipart>=0.0.9
spacy>=3.7.4
requests>=2.31.0
slowapi>=0.1.9
reportlab>=4.1.0
openpyxl>=3.1.2
scipy>=1.13.0
